refactor: requested changes
