+++ interactions/api/http/interaction.py
@@ -2,15 +2,15 @@ from typing import List, Optional, Union
 
 from ...api.cache import Cache
 from ..models import Snowflake
-from .Request import Request
-from .Route import Route
+from .request import _Request
+from .route import Route
 
 
-class HTTPInteraction:
+class _InteractionRequest:
 
     __slots__ = ("_req", "cache")
 
-    _req: Request
+    _req: _Request
     cache: Cache
 
     def __init__(self, _req, cache) -> None:
