+++ interactions/api/http/scheduledEvent.pyi
@@ -3,14 +3,14 @@ from typing import List
 
 from ...api.cache import Cache
 from ..models import Snowflake
-from .Request import Request
+from .request import _Request
 
 
-class HTTPScheduledEvent:
+class _ScheduledEventRequest:
 
     __slots__ = ("_req", "cache")
 
-    _req: Request
+    _req: _Request
     cache: Cache
 
     def __init__(self, _req, cache) -> None: ...
