+++ interactions/api/http/sticker.py
@@ -3,15 +3,15 @@ from typing import List, Optional
 from aiohttp import FormData
 
 from ...api.cache import Cache
-from .Request import Request
-from .Route import Route
+from .request import _Request
+from .route import Route
 
 
-class HTTPSticker:
+class _StickerRequest:
 
     __slots__ = ("_req", "cache")
 
-    _req: Request
+    _req: _Request
     cache: Cache
 
     def __init__(self, _req, cache) -> None:
