+++ interactions/api/http/sticker.pyi
@@ -3,12 +3,12 @@ from typing import List, Optional
 from aiohttp import FormData
 
 from ...api.cache import Cache
-from .Request import Request
+from .request import _Request
 
 
-class HTTPSticker:
+class _StickerRequest:
 
-    _req: Request
+    _req: _Request
     cache: Cache
 
     def __init__(self, _req, cache) -> None: ...
