+++ interactions/api/http/thread.pyi
@@ -1,11 +1,11 @@
 from typing import List, Optional
 
 from ...api.cache import Cache
-from .Request import Request
+from .request import _Request
 
-class HTTPThread:
+class _ThreadRequest:
 
-    _req: Request
+    _req: _Request
     cache: Cache
 
     def __init__(self, _req, cache) -> None: ...
