+++ interactions/api/http/webhook.pyi
@@ -1,12 +1,12 @@
 from typing import Any, List, Optional
 
 from ...api.cache import Cache
-from .Request import Request
+from .request import _Request
 
 
-class HTTPWebhook:
+class _WebhookRequest:
 
-    _req: Request
+    _req: _Request
     cache: Cache
 
     def __init__(self, _req, cache) -> None: ...
